Rework of measurements
Now always meters (for lengths) or square meters (for areas) are returned
and the geometry is always returned in lat/lon coordinates.
Furthermore the lengths and areas are calculated geodesic, which adjusts
for the underlying sphere/geoid.

The sample has been changed accordingly.

The deprecated functions in MeasureEvent might be removed later on.
They just affect getting measurements using a different projection,
which does not make anymore sense.