Prepare for sonatype releases
It's important to use SNAPSHOT version so the project not get deployed as a release, used the HEAD-SNAPSHOT strategy which is easier and people using SNAPSHOT version do not need to update the SNAPSHOT version, the will keep in the head forever. This also go perfect with the creative versioning used here ;).
Renamed parent module to gwt-ol3-parent and client from gwt-ol3-client to gwt-ol3 which is the actual lib name.
Collapsed permutations, OL3 is a one script lib, and the demo size is just 90k, after collapsing all permutation it get up to 100k which is almost the same. So collapsed.